chore: UsuarioService and more

+++ src/main/java/com/example/demo/DTO/request/MovimientoCreacionDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MovimientoCreacionDTO {
-    private int id_instructor;
+    private Long id_instructor;
     private LocalDate fecha;
     private TipoMovimiento tipo;
 }
